Refactored Code / Ready to use
It is working, I changed the stuff that I saw that could be better. I know there are more to change. But I tested, works fine.
